refresh button - initial functionality
Added a button above the NFT image that forces a refetching of the metadata. I also had to (I think...) make the tokenURI variable non-const (var) so that it can be changed from this function. The added function forces a refetch of the contract data to get the URI, uses axios to get the JSON from that URI, and then compares it to the stored data.metadata. If they are the same, do nothing, if they are different, update data.metadata, data.token_uri, and tokenURI.

Need to probably add some push to the backend api to store the new data, but I'm not sure.

I also suspect the getJsonFromURI function is not doing what it is thought to be doing. it parses a string into a JSON, but usually a URL is being passed in. I think it's supposed to go get the json and parse it and return it.