fixing linter
mostly fixing linter issues and updated NFTItemType to have metadata and token_uri, since clearly that's in the database. The types need to be aligned with what's being stored in the DB. I can't find where the data is getting pushed, but it's some mix of the Nft model plus other data. I also changed the fetchNFTDetail to not return an array of NFTItemType since it's grabbing one token data. it doesn't make sense to return an array here. Also added check for IPFS metadata and will pull that instead of expecting it to be an https